requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
PyMySQL>=1.0.0
gunicorn>=20.1.0
redis>=4.5.0
django-redis>=5.2.0
celery>=5.2.0
Pillow>=10.0.0
reportlab>=4.0.0
xlsxwriter>=3.1.0
pytz>=2023.3
python-dateutil>=2.8.0
